requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas==2.1.4
gspread==5.12.0
google-generativeai==0.3.2
openpyxl==3.1.2
